pkcs15-crypt add use-pinpad option
Added an use-pinpad option similar to pkcs15-init.c.
A usecase is to boot a headless server with a LUKS encrypted /var/lib-partition and a pinpad reader, eg ReinerSCT with aventra MyEID. Therfore your backups from eg KDC, LDAP are highly secured. The ReinerSCT works also in a Virtual Machine with USB-Passthrough on ESXi 6.7.